Move ImageVector method parsing to ComposeMethodCreator class
